refactor: Sections 객체 구현
